QEventDispatcherCF: avoid QFlags->int implicit casts
Use toInt(), or an explicit cast. (I don't quite know why an atomic int
is involved here, but anyways, it makes these casts necessary.)

+++ src/corelib/kernel/qeventdispatcher_cf_p.h
@@ -236,7 +236,7 @@ protected:
     struct ProcessEventsState
     {
         ProcessEventsState(QEventLoop::ProcessEventsFlags f)
-         : flags(f), wasInterrupted(false)
+         : flags(f.toInt()), wasInterrupted(false)
          , processedPostedEvents(false), processedTimers(false)
          , deferredWakeUp(false), deferredUpdateTimers(false) {}
